backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
aiosqlite==0.21.0
httpx==0.28.1
feedparser==6.0.11
rapidfuzz==3.14.1
python-dateutil==2.9.0.post0
apscheduler==3.11.0
pydantic-settings==2.10.1
beautifulsoup4==4.13.4
